@@ -4,67 +4,76 @@
   "license": "MIT",
   "angular-cli": {},
   "scripts": {
-    "start": "npm run build-themes && ng serve",
-    "start-jit": "npm run build-themes && ng serve --aot=false",
-    "lint": "tslint -p src/tsconfig.json",
-    "test": "npm run build-themes && ng test",
+    "ng": "ng",
+    "start": "yarn build-themes && ng serve",
+    "start:jit": "yarn build-themes && ng serve --aot=false",
+    "start:prod": "yarn build-themes && ng serve --prod",
+    "lint": "ng lint",
+    "test": "yarn build-themes && ng test",
     "pree2e": "webdriver-manager update",
-    "e2e": "protractor",
-    "fetch-local": "bash ./tools/fetch-assets-local.sh",
+    "e2e": "yarn build-themes && ng e2e",
+    "build": "yarn build-themes && ng build",
+    "build:content": "yarn upgrade @angular/components-examples",
+    "build:highmem": "yarn build-themes && node --max_old_space_size=8192 ./node_modules/@angular/cli/bin/ng build",
+    "build:sm": "ng build --prod --source-map",
     "build-themes": "bash ./tools/build-themes.sh",
-    "prod-build": "npm run build-themes && ng build --aot --prod",
-    "postinstall": "webdriver-manager update && bash ./tools/build-themes.sh",
+    "prod-build": "yarn build-themes && ng build --prod",
+    "postinstall": "webdriver-manager update --gecko false && yarn build-themes && ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points",
     "publish-prod": "bash ./tools/deploy.sh stable prod",
     "publish-dev": "bash ./tools/deploy.sh",
     "publish-beta": "bash ./tools/deploy.sh stable beta"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.0.0",
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.0.0",
-    "@angular/compiler": "^7.0.0",
-    "@angular/core": "^7.0.0",
-    "@angular/forms": "^7.0.0",
-    "@angular/http": "^7.0.0",
-    "@angular/material": "^7.2.0",
-    "@angular/material-moment-adapter": "^7.2.0",
-    "@angular/platform-browser": "^7.0.0",
-    "@angular/platform-browser-dynamic": "^7.0.0",
-    "@angular/router": "^7.0.0",
-    "core-js": "^2.5.6",
-    "hammerjs": "^2.0.8",
-    "moment": "^2.22.1",
-    "rxjs": "^6.1.0",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "9.0.0",
+    "@angular/cdk": "^9.0.1",
+    "@angular/cdk-experimental": "^9.0.1",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/components-examples": "angular/material2-docs-content#9.0.x",
+    "@angular/core": "9.0.0",
+    "@angular/forms": "9.0.0",
+    "@angular/google-maps": "^9.0.1",
+    "@angular/material": "^9.0.1",
+    "@angular/material-experimental": "^9.0.1",
+    "@angular/material-moment-adapter": "^9.0.1",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@angular/youtube-player": "^9.0.1",
+    "core-js": "^2.6.11",
+    "material-components-web": "^4.0.0",
+    "moment": "^2.24.0",
+    "rxjs": "^6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.10.1",
-    "@angular/cli": "^7.0.1",
-    "@angular/compiler-cli": "^7.0.0",
+    "@angular-devkit/build-angular": "0.900.0",
+    "@angular/cli": "9.0.0",
+    "@angular/compiler-cli": "9.0.0",
+    "@angular/localize": "9.0.0",
     "@awesome-fe/translate": "^0.3.2",
-    "@types/jasmine": "^2.8.9",
-    "@types/node": "^10.0.9",
-    "firebase-tools": "^6.0.0",
+    "@types/jasmine": "^3.5.2",
+    "@types/node": "^12.12.24",
+    "codelyzer": "^5.2.1",
+    "firebase-tools": "^7.12.1",
     "gulp": "^3.9.1",
     "gulp-replace": "^1.0.0",
     "hammerjs": "^2.0.8",
     "highlight.js": "^9.9.0",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-browserstack-launcher": "^1.3.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-sauce-launcher": "^1.2.0",
-    "natives": "^1.1.6",
-    "node-sass": "^4.5.3",
-    "protractor": "~5.3.0",
-    "ts-node": "^6.0.3",
-    "tslint": "^5.10.0",
-    "typescript": "^3.1.3"
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-firefox-launcher": "^1.3.0",
+    "karma-jasmine": "^3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "node-sass": "^4.13.1",
+    "protractor": "^5.4.3",
+    "ts-node": "8.6.2",
+    "tslint": "^5.20.1",
+    "typescript": "3.6.4"
   }
 }
