@@ -1,48 +1,48 @@
 import {
-  Component,
-  ViewEncapsulation,
   ChangeDetectionStrategy,
+  Component,
   NgModule,
-  OnInit,
   OnDestroy,
+  OnInit,
+  ViewEncapsulation,
 } from '@angular/core';
-import {StyleManager} from '../style-manager/style-manager';
-import {ThemeStorage, DocsSiteTheme} from './theme-storage/theme-storage';
-import {
-  MatButtonModule,
-  MatGridListModule,
-  MatIconModule,
-  MatMenuModule,
-  MatTooltipModule,
-} from '@angular/material';
+import {StyleManager} from '../style-manager';
+import {DocsSiteTheme, ThemeStorage} from './theme-storage/theme-storage';
+import {MatButtonModule} from '@angular/material/button';
+import {MatIconModule, MatIconRegistry} from '@angular/material/icon';
+import {MatMenuModule} from '@angular/material/menu';
+import {MatTooltipModule} from '@angular/material/tooltip';
 import {CommonModule} from '@angular/common';
-import {ActivatedRoute} from '@angular/router';
+import {ActivatedRoute, ParamMap} from '@angular/router';
 import {Subscription} from 'rxjs';
-import {map, filter} from 'rxjs/operators';
-
+import {map} from 'rxjs/operators';
+import {DomSanitizer} from '@angular/platform-browser';
+import {LiveAnnouncer} from '@angular/cdk/a11y';
 
 @Component({
   selector: 'theme-picker',
   templateUrl: 'theme-picker.html',
   styleUrls: ['theme-picker.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  encapsulation: ViewEncapsulation.None,
-  host: {'aria-hidden': 'true'},
+  encapsulation: ViewEncapsulation.None
 })
 export class ThemePicker implements OnInit, OnDestroy {
   private _queryParamSubscription = Subscription.EMPTY;
   currentTheme: DocsSiteTheme;
 
+  // The below colors need to align with the themes defined in theme-picker.scss
   themes: DocsSiteTheme[] = [
     {
       primary: '#673AB7',
       accent: '#FFC107',
+      displayName: 'Deep Purple & Amber',
       name: 'deeppurple-amber',
       isDark: false,
     },
     {
       primary: '#3F51B5',
       accent: '#E91E63',
+      displayName: 'Indigo & Pink',
       name: 'indigo-pink',
       isDark: false,
       isDefault: true,
@@ -50,35 +50,49 @@ export class ThemePicker implements OnInit, OnDestroy {
     {
       primary: '#E91E63',
       accent: '#607D8B',
+      displayName: 'Pink & Blue-grey',
       name: 'pink-bluegrey',
       isDark: true,
     },
     {
       primary: '#9C27B0',
       accent: '#4CAF50',
+      displayName: 'Purple & Green',
       name: 'purple-green',
       isDark: true,
     },
   ];
 
-  constructor(
-    public styleManager: StyleManager,
-    private _themeStorage: ThemeStorage,
-    private _activatedRoute: ActivatedRoute) {
-    this.installTheme(this._themeStorage.getStoredThemeName());
+  constructor(public styleManager: StyleManager,
+              private _themeStorage: ThemeStorage,
+              private _activatedRoute: ActivatedRoute,
+              private liveAnnouncer: LiveAnnouncer,
+              iconRegistry: MatIconRegistry,
+              sanitizer: DomSanitizer) {
+    iconRegistry.addSvgIcon('theme-example',
+                            sanitizer.bypassSecurityTrustResourceUrl(
+                                'assets/img/theme-demo-icon.svg'));
+    const themeName = this._themeStorage.getStoredThemeName();
+    if (themeName) {
+      this.selectTheme(themeName);
+    }
   }
 
   ngOnInit() {
     this._queryParamSubscription = this._activatedRoute.queryParamMap
-      .pipe(map(params => params.get('theme')), filter(Boolean))
-      .subscribe(themeName => this.installTheme(themeName));
+      .pipe(map((params: ParamMap) => params.get('theme')))
+      .subscribe((themeName: string | null) => {
+        if (themeName) {
+          this.selectTheme(themeName);
+        }
+    });
   }
 
   ngOnDestroy() {
     this._queryParamSubscription.unsubscribe();
   }
 
-  installTheme(themeName: string) {
+  selectTheme(themeName: string) {
     const theme = this.themes.find(currentTheme => currentTheme.name === themeName);
 
     if (!theme) {
@@ -94,6 +108,7 @@ export class ThemePicker implements OnInit, OnDestroy {
     }
 
     if (this.currentTheme) {
+      this.liveAnnouncer.announce(`${theme.displayName} theme selected.`, 'polite', 3000);
       this._themeStorage.storeTheme(this.currentTheme);
     }
   }
@@ -101,12 +116,11 @@ export class ThemePicker implements OnInit, OnDestroy {
 
 @NgModule({
   imports: [
+    CommonModule,
     MatButtonModule,
     MatIconModule,
     MatMenuModule,
-    MatGridListModule,
     MatTooltipModule,
-    CommonModule
   ],
   exports: [ThemePicker],
   declarations: [ThemePicker],
