@@ -1,11 +1,12 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {MatSidenav} from '@angular/material';
 import {take} from 'rxjs/operators';
 import {ComponentSidenav, ComponentSidenavModule} from './component-sidenav';
 import {DocsAppTestingModule} from '../../testing/testing-module';
+import {MatSidenav} from '@angular/material/sidenav';
 
 describe('ComponentSidenav', () => {
   let fixture: ComponentFixture<ComponentSidenav>;
+  let component: ComponentSidenav;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -15,17 +16,17 @@ describe('ComponentSidenav', () => {
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ComponentSidenav);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
   });
 
   it('should close the sidenav on init', () => {
-    const component = fixture.componentInstance;
-
     // Spy on window.mediaMatch and return stub
     spyOn(window, 'matchMedia').and.returnValue({
       matches: true
-    });
+    } as any);
 
-    fixture.detectChanges();
+    // TODO refactor this as none of these expectations are ever verified
     async(() => {
       expect(component.sidenav instanceof MatSidenav).toBeTruthy();
       component.isScreenSmall.pipe(take(1)).subscribe(isSmall => expect(isSmall).toBeTruthy());
@@ -34,10 +35,6 @@ describe('ComponentSidenav', () => {
   });
 
   it('should show a link for each item in doc items categories', () => {
-    const component = fixture.componentInstance;
-
-    fixture.detectChanges();
-
     const totalItems = component.docItems.getItems('categories').length;
     const totalLinks = fixture
       .nativeElement
