@@ -1,10 +1,11 @@
 import {Component, NgModule, OnInit} from '@angular/core';
-import {ActivatedRoute, RouterModule, Router} from '@angular/router';
+import {ActivatedRoute, Router, RouterModule, Routes} from '@angular/router';
 import {GuideItem, GuideItems} from '../../shared/guide-items/guide-items';
 import {FooterModule} from '../../shared/footer/footer';
 import {DocViewerModule} from '../../shared/doc-viewer/doc-viewer-module';
 import {TableOfContentsModule} from '../../shared/table-of-contents/table-of-contents.module';
 import {ComponentPageTitle} from '../page-title/page-title';
+import {ReactiveFormsModule} from '@angular/forms';
 
 @Component({
   selector: 'guide-viewer',
@@ -12,14 +13,17 @@ import {ComponentPageTitle} from '../page-title/page-title';
   styleUrls: ['./guide-viewer.scss'],
 })
 export class GuideViewer implements OnInit {
-  guide: GuideItem;
+  guide: GuideItem | undefined;
 
   constructor(_route: ActivatedRoute,
               private _componentPageTitle: ComponentPageTitle,
               private router: Router,
               public guideItems: GuideItems) {
     _route.params.subscribe(p => {
-      this.guide = guideItems.getItemById(p['id']);
+      const guideItem = guideItems.getItemById(p['id']);
+      if (guideItem) {
+        this.guide = guideItem;
+      }
 
       if (!this.guide) {
         this.router.navigate(['/guides']);
@@ -28,14 +32,24 @@ export class GuideViewer implements OnInit {
   }
 
   ngOnInit(): void {
-    this._componentPageTitle.titleCn = this.guide.name;
+    if (this.guide !== undefined) {
+      this._componentPageTitle.titleCn = this.guide.name;
+    }
   }
 }
 
+const routes: Routes = [ {path : '', component : GuideViewer} ];
+
+// This module needs to include all of the modules required by the examples in the guides.
+// For example, the custom form-field guide requires the ReactiveFormsModule.
+// These imports may need to be updated when adding examples to new or existing guides.
 @NgModule({
-  imports: [DocViewerModule, FooterModule, RouterModule, TableOfContentsModule],
+  imports: [
+    DocViewerModule, FooterModule, TableOfContentsModule, ReactiveFormsModule,
+    RouterModule.forChild(routes),
+  ],
   exports: [GuideViewer],
   declarations: [GuideViewer],
-  providers: [GuideItems, ComponentPageTitle],
+  providers: [GuideItems],
 })
 export class GuideViewerModule {}
