@@ -1,8 +1,8 @@
 import {Component, NgModule, OnInit} from '@angular/core';
 import {SvgViewerModule} from '../../shared/svg-viewer/svg-viewer';
-import {MatButtonModule} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
 import {FooterModule} from '../../shared/footer/footer';
-import {RouterModule} from '@angular/router';
+import {RouterModule, Routes} from '@angular/router';
 import {ComponentPageTitle} from '../page-title/page-title';
 
 @Component({
@@ -21,8 +21,10 @@ export class Homepage implements OnInit {
   }
 }
 
+const routes: Routes = [ {path : '', component : Homepage} ];
+
 @NgModule({
-  imports: [SvgViewerModule, MatButtonModule, FooterModule, RouterModule],
+  imports: [SvgViewerModule, MatButtonModule, FooterModule, RouterModule.forChild(routes)],
   exports: [Homepage],
   declarations: [Homepage],
 })
