@@ -4,9 +4,10 @@ import { Title } from '@angular/platform-browser';
 /**
  * Service responsible for setting the title that appears above the components and guide pages.
  */
-@Injectable()
+@Injectable({providedIn: 'root'})
 export class ComponentPageTitle {
   _title = '';
+  _originalTitle = 'Angular Material UI component library';
 
   get title(): string {
     return this._title;
@@ -15,7 +16,9 @@ export class ComponentPageTitle {
   set title(title: string) {
     this._title = title;
     if (title !== '') {
-      title = `${title} | `;
+      title = `${title} | Angular Material`;
+    } else {
+      title = this._originalTitle;
     }
     this.updateTitle();
   }
@@ -35,10 +38,9 @@ export class ComponentPageTitle {
   }
 
   private updateTitle() {
-    this.bodyTitle.setTitle(`${this.titleCn} ${this.title} - Angular Material`);
+    this.bodyTitle.setTitle(`${this.titleCn} ${this.title}`);
   }
 
   constructor(private bodyTitle: Title) {
   }
-
 }
