@@ -1,10 +1,14 @@
 import {Component, NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {MatButtonModule, MatMenuModule} from '@angular/material';
+import {MatButtonModule} from '@angular/material/button';
+import {MatMenuModule} from '@angular/material/menu';
 import {RouterModule} from '@angular/router';
-import {ThemePickerModule} from '../theme-picker/theme-picker';
+import {ThemePickerModule} from '../theme-picker';
 import {VersionPickerModule} from '../version-picker';
 import {SECTIONS} from '../documentation-items/documentation-items';
+import {ThemeStorage} from '../theme-picker/theme-storage/theme-storage';
+import {StyleManager} from '../style-manager';
+import {HttpClientModule} from '@angular/common/http';
 
 const SECTIONS_KEYS = Object.keys(SECTIONS);
 
@@ -27,14 +31,16 @@ export class NavBar {
 
 @NgModule({
   imports: [
+    CommonModule,
+    HttpClientModule,
     MatButtonModule,
     MatMenuModule,
     RouterModule,
     ThemePickerModule,
     VersionPickerModule,
-    CommonModule
   ],
   exports: [NavBar],
   declarations: [NavBar],
+  providers: [StyleManager, ThemeStorage]
 })
 export class NavBarModule {}
