@@ -9,10 +9,11 @@ import {unregisterServiceWorkers} from './unregister-service-workers';
 // Unregister all installed service workers and force reload the page if there was
 // an old service worker from a previous version of the docs.
 unregisterServiceWorkers()
-  .then(hadServiceWorker => hadServiceWorker && location.reload(true));
+  .then(hadServiceWorker => hadServiceWorker && location.reload());
 
 if (environment.production) {
   enableProdMode();
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule);
+platformBrowserDynamic().bootstrapModule(AppModule)
+  .catch(err => console.error(err));
